move files to DetailedPage folder
