+++ src/components/modules/DetailedPage/PositionCard/PositionCard.tsx
