+++ src/components/modules/DetailedPage/PositionCard/StyledPositionCard.tsx
