+++ src/components/modules/DetailedPage/PositionCard/index.ts
