+++ src/components/modules/DetailedPage/ProductHeader/index.js
