+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import { GetStaticProps } from "next";
 import Head from "next/head";
-import PositionCard from "../components/modules/IdeaRecruitPage/PositionCard";
+import PositionCard from "../components/modules/DetailedPage/PositionCard";
 import FilteringComponent from "../components/modules/MainPage/filtering/FilteringComponent";
 import { Project } from "../components/modules/MainPage/filtering/project";
